Adds the first version of the ViewController for data input
This first version uses RxSwift to react to changes in the data and showing the results.

+++ Clients/Apple/FitnessTracker/FitnessTracker/FitnessInfo.swift
@@ -8,7 +8,7 @@
 
 import Foundation
 
-typealias Weight = UInt
+typealias Weight = Double
 typealias Height = UInt
 
 public struct FitnessInfo {
